simple csv extractor example
For the CDF Bootcamp we need a simple csv extractor example that could be deployed as a Cognite Function. 

This PR contains a simplified version of the csv-extractor sample that reads CSV files and uploads the content to a configurable
database and table in CDF RAW. This extractor can be deployed as a Cognite Function. 
This simple version differ from the csv-extractor sample by
1. No state store
2. No Prometheus metrics
3. Added extraction pipeline in config
4. The config file path can be passed into extractor.main() so it can be retrieved from the data JSON when running as a Cognite Function